+++ src/main/java/org/ohdsi/webapi/security/model/EntityPermissionSchema.java
@@ -1,4 +1,4 @@
-package org.ohdsi.webapi.security.entity;
+package org.ohdsi.webapi.security.model;
 
 import java.util.Collections;
 import java.util.Map;
