+++ src/main/java/org/ohdsi/webapi/shiro/management/AtlasSecurity.java
@@ -21,8 +21,8 @@ import org.apache.shiro.web.servlet.AdviceFilter;
 import org.apache.shiro.web.util.WebUtils;
 import org.ohdsi.webapi.OidcConfCreator;
 import org.ohdsi.webapi.cohortcharacterization.CcImportEvent;
-import org.ohdsi.webapi.security.entity.EntityPermissionSchemaResolver;
-import org.ohdsi.webapi.security.entity.EntityType;
+import org.ohdsi.webapi.security.model.EntityPermissionSchemaResolver;
+import org.ohdsi.webapi.security.model.EntityType;
 import org.ohdsi.webapi.shiro.Entities.RoleEntity;
 import org.ohdsi.webapi.shiro.PermissionManager;
 import org.ohdsi.webapi.shiro.filters.CorsFilter;
